Backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
typing_inspect
fastapi-utils
passlib
bcrypt==4.0.1
python-multipart
"python-jose[cryptography]"